feat: 일치 결과를 WinningLotto 객체에서 계산하도록 변경
