+++ src/main/kotlin/lotto/domain/WinningLotto.kt
@@ -5,6 +5,22 @@ class WinningLotto(val lottoNumbers: LottoNumbers) {
         return lottoNumbers.getNumbers()
     }
 
+    fun match(lottos: List<Lotto>): LottoResult {
+        val lottoResult = LottoResult()
+        lottos.forEach { lotto ->
+            val lottoNumber = lotto.getNumbers()
+            var count = 0
+            lottoNumber.forEach { number ->
+                if (getNumbers().contains(number)) {
+                    count++
+                }
+            }
+
+            lottoResult.setLottoResult(count)
+        }
+        return lottoResult
+    }
+
     companion object {
         fun create(numbers: List<Int>): WinningLotto {
             val lottoNumbers = LottoNumbers(numbers.map { LottoNumber(it) })
